Added "success" messages to the CGI pages. We have error messages, but not many success messages. In most cases, due to re-direct, you won't see them. But, I'm going to make it an option (for "headers already sent" errors.)

+++ scripts/sb_fileio.php
@@ -173,7 +173,11 @@
 		//
 		// I'm guessing that we don't need to lock a file to delete it. Who knows?
 		clearstatcache();
-		$result = @unlink( $filename );
+		if ( file_exists( $filename ) ) {
+			$result = @unlink( $filename );
+		} else {
+			$result = 'File "' . $filename . '" does not exist.';
+		}
 		return( $result );
 	}
